+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 7.43
 -----
+- The search bar on the podcasts tab is now visible by default [#929]
 - Correctly display chapter durations when the chapter info shows that it passes the end of the episode.
 
 7.42